+++ packages/apps-config/src/api/spec/index.ts
@@ -9,6 +9,7 @@ import aresParachain from './ares-parachain';
 import bifrost from './bifrost';
 import bifrostParachain from './bifrost-parachain';
 import bitcountry from './bitcountry';
+import bitcountryParachain from './bitcountry-rococo';
 import canvas from './canvas';
 import centrifugeChain from './centrifuge-chain';
 import chainx from './chainx';
@@ -30,20 +31,25 @@ import encointerNodeTeeproxy from './encointer-node-teeproxy';
 import encointerPara from './encointer-para';
 import equilibrium from './equilibrium';
 import galital from './galital';
+import galitalParachain from './galital-parachain';
 import galois from './galois';
+import gamepower from './gamepower';
 import hanonycash from './hanonycash';
 import hydrate from './hydrate';
 import idavoll from './idavoll';
 import integritee from './integritee';
+import ipse from './ipse';
 import jupiter from './jupiter';
 import jupiterRococo from './jupiter-rococo';
 import kilt from './kilt';
+import konomi from './konomi';
 import kulupu from './kulupu';
 import laminar from './laminar';
 import litentry from './litentry';
 import manta from './manta';
 import moonbeam from './moonbeam';
 import mybank from './mybank';
+import neatcoin from './neatcoin';
 import nftmart from './nftmart';
 import nodle from './nodle';
 import parami from './parami';
@@ -55,11 +61,11 @@ import plasmParachain from './plasm-parachain';
 import polkabtc from './polkabtc';
 import polkadex from './polkadex';
 import polkafoundry from './polkafoundry';
+import polymesh from './polymesh';
 import prism from './prism';
 import realis from './realis';
 import riochain from './riochain';
 import robonomics from './robonomics';
-import sgc from './sgc';
 import snowbridge from './snowbridge';
 import soraSubstrate from './soraSubstrate';
 import stafi from './stafi';
@@ -72,6 +78,8 @@ import unique from './unique';
 import unitv from './unitv';
 import vln from './vln';
 import vlnrococo from './vln-rococo';
+import vodka from './vodka';
+import web3games from './web3games';
 import westlake from './westlake';
 import zeitgeist from './zeitgeist';
 import zenlink from './zenlink';
@@ -92,6 +100,7 @@ const spec: Record<string, OverrideBundleDefinition> = {
   bifrost: bifrost,
   'bifrost-parachain': bifrostParachain,
   'bitcountry-node': bitcountry,
+  'bitcountry-parachain': bitcountryParachain,
   'btc-parachain': polkabtc,
   canvas,
   'centrifuge-chain': centrifugeChain,
@@ -114,15 +123,20 @@ const spec: Record<string, OverrideBundleDefinition> = {
   'encointer-node-notee': encointerNodeNotee,
   'encointer-node-teeproxy': encointerNodeTeeproxy,
   'encointer-parachain': encointerPara,
-  'galital-collator': galital,
+  galital: galital,
+  'galital-collator': galitalParachain,
+  gamepower,
   'hack-hydra-dx': hydrate,
+  halongbay: polkafoundry,
   hanonycash,
   'hydra-dx': hydrate,
   idavoll,
   'integritee-parachain': integritee,
+  'ipse-node': ipse,
   'jupiter-prep': jupiter,
   'jupiter-rococo': jupiterRococo,
   'kilt-parachain': kilt,
+  konomi,
   kulupu,
   laminar,
   litentry,
@@ -134,6 +148,7 @@ const spec: Record<string, OverrideBundleDefinition> = {
   moonbeam,
   'moonbeam-standalone': moonbeam,
   'mybank.network Testnet': mybank,
+  neatcoin,
   nft: unique,
   nftmart,
   'node-moonbeam': moonbeam,
@@ -146,12 +161,11 @@ const spec: Record<string, OverrideBundleDefinition> = {
   'phoenix-parachain': phoenix,
   plasm,
   'plasm-parachain': plasmParachain,
-  polkafoundry,
+  polymesh,
   prism,
   realis,
   'riochain-runtime': riochain,
   robonomics,
-  'sgc-node': sgc,
   snowbridge,
   'sora-substrate': soraSubstrate,
   stafi,
@@ -164,6 +178,8 @@ const spec: Record<string, OverrideBundleDefinition> = {
   uart,
   'unit-node': unitv,
   'unit-parachain': unitv,
+  vodka,
+  'web3games-node': web3games,
   westlake: westlake,
   zeitgeist: zeitgeist
 };