+++ packages/apps-config/src/endpoints/testing.ts
@@ -61,6 +61,7 @@ export function createTesting (t: TFunction): LinkOption[] {
     },
     {
       info: 'clover',
+      isDisabled: true, // Cannot construct unknown type BridgeNetworks
       text: t('rpc.clover.finance', 'Clover', { ns: 'apps-config' }),
       providers: {
         Clover: 'wss://api.clover.finance/'
@@ -94,13 +95,6 @@ export function createTesting (t: TFunction): LinkOption[] {
       text: t('rpc.dotmog', 'DOTMog', { ns: 'apps-config' }),
       providers: {
         DOTMog: 'wss://mogiway-01.dotmog.com'
-      }
-    },
-    {
-      info: 'phoenix',
-      text: t('rpc.phoenix', 'Phoenix Mashnet', { ns: 'apps-config' }),
-      providers: {
-        'phoenix Protocol': 'wss://phoenix-ws.coinid.pro/'
       }
     },
     {
@@ -139,6 +133,13 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
+      info: 'Galital',
+      text: t('rpc.galital', 'Galital PC2', { ns: 'apps-config' }),
+      providers: {
+        StarkleyTech: 'wss://galital-rpc-testnet.starkleytech.com'
+      }
+    },
+    {
       info: 'galois',
       text: t('rpc.galois', 'Galois', { ns: 'apps-config' }),
       providers: {
@@ -147,19 +148,34 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
+      info: 'gamepower',
+      text: t('rpc.gamepower', 'GamePower', { ns: 'apps-config' }),
+      providers: {
+        GamePower: 'wss://gamepower.io'
+      }
+    },
+    {
+      info: 'halongbay',
+      text: 'Halongbay Testnet',
+      providers: {
+        Halongbay: 'wss://halongbay.polkafoundry.com'
+      }
+    },
+    {
+      info: 'ipse',
+      text: t('rpc.ipse', 'IPSE', { ns: 'apps-config' }),
+      providers: {
+        'IPSE China': 'wss://testnet-china.ipse.io',
+        'IPSE USA': 'wss://testnet-usa.ipse.io',
+        'IPSE Europe': 'wss://testnet-europe.ipse.io'
+      }
+    },
+    {
       info: 'jupiter',
       text: t('rpc.jupiter', 'Jupiter', { ns: 'apps-config' }),
       providers: {
         Elara: 'wss://jupiter-poa.elara.patract.io',
         Patract: 'wss://ws.jupiter-poa.patract.cn'
-      }
-    },
-    {
-      info: 'acala',
-      text: t('rpc.mandala', 'Mandala', { ns: 'apps-config' }),
-      providers: {
-        Acala: 'wss://acala-mandala.api.onfinality.io/public-ws',
-        'Patract Elara': 'wss://mandala.elara.patract.io'
       }
     },
     {
@@ -177,18 +193,49 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
+      info: 'acala',
+      text: t('rpc.mandala', 'Mandala', { ns: 'apps-config' }),
+      providers: {
+        Acala: 'wss://acala-mandala.api.onfinality.io/public-ws',
+        'Patract Elara': 'wss://mandala.elara.patract.io'
+      }
+    },
+    {
       info: 'moonbaseAlpha',
       text: t('rpc.moonbeam', 'Moonbase Alpha', { ns: 'apps-config' }),
       providers: {
         'Moonbeam Network': 'wss://wss.testnet.moonbeam.network',
-        OnFinality: 'wss://moonbeam-alpha.api.onfinality.io/public-ws'
+        OnFinality: 'wss://moonbeam-alpha.api.onfinality.io/public-ws',
+        'Patract Elara': 'wss://moonbase.moonbeam.elara.patract.io'
+      }
+    },
+    {
+      info: 'mybank',
+      text: t('rpc.mybank', 'mybank.network', { ns: 'apps-config' }),
+      providers: {
+        MYBANK: 'wss://mybank.network/substrate'
+      }
+    },
+    {
+      info: 'nftmart',
+      text: t('rpc.nftmart', 'NFTMart', { ns: 'apps-config' }),
+      providers: {
+        NFTMartDev: 'wss://dev-ws.nftmart.io',
+        NFTMartStaging: 'wss://staging-ws.nftmart.io'
       }
     },
     {
       info: 'phala',
-      text: t('rpc.phala', 'Phala PoC-3', { ns: 'apps-config' }),
+      text: t('rpc.phala', 'Phala PoC-4', { ns: 'apps-config' }),
       providers: {
-        'Phala Network': 'wss://poc3a.phala.network/ws'
+        'Phala Network': 'wss://poc4.phala.network/ws'
+      }
+    },
+    {
+      info: 'phoenix',
+      text: t('rpc.phoenix', 'Phoenix Mashnet', { ns: 'apps-config' }),
+      providers: {
+        'phoenix Protocol': 'wss://phoenix-ws.coinid.pro/'
       }
     },
     {
@@ -203,6 +250,13 @@ export function createTesting (t: TFunction): LinkOption[] {
       text: t('rpc.polkabtc', 'PolkaBTC', { ns: 'apps-config' }),
       providers: {
         Interlay: 'wss://beta.polkabtc.io/api/parachain'
+      }
+    },
+    {
+      info: 'polymesh',
+      text: t('rpc.polymesh', 'Polymesh ITN', { ns: 'apps-config' }),
+      providers: {
+        Polymath: 'wss://itn-rpc.polymesh.live'
       }
     },
     {
@@ -227,15 +281,8 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
-      info: 'sgc',
-      text: t('rpc.sgc', 'Sgc', { ns: 'apps-config' }),
-      providers: {
-        SGC: 'wss://substrate.org.cn:4443'
-      }
-    },
-    {
       info: 'sora-substrate',
-      text: t('rpc.sora-substrate', 'SORA-staging', { ns: 'apps-config' }),
+      text: t('rpc.sora-substrate-staging', 'SORA-staging', { ns: 'apps-config' }),
       providers: {
         Soramitsu: 'wss://ws.stage.sora2.soramitsu.co.jp'
       }
@@ -256,9 +303,9 @@ export function createTesting (t: TFunction): LinkOption[] {
     },
     {
       info: 'uniarts',
-      text: t('rpc.uniarts', 'Uniarts', { ns: 'apps-config' }),
+      text: t('rpc.uniarts', 'UniArts', { ns: 'apps-config' }),
       providers: {
-        Uniarts: 'wss://testnet.uniarts.me'
+        UniArts: 'wss://testnet.uniarts.me'
       }
     },
     {
@@ -269,19 +316,24 @@ export function createTesting (t: TFunction): LinkOption[] {
       }
     },
     {
-      info: 'vln',
-      text: t('rpc.vln', 'Valiu Liquidity Network', { ns: 'apps-config' }),
+      info: 'unitv',
+      text: t('rpc.unitv', 'Unit Network', { ns: 'apps-config' }),
       providers: {
-        Valiu: 'wss://vln.valiu.dev'
+        'Unit Network': 'wss://unitventures.io/'
       }
     },
     {
-      dnslink: 'westend',
-      info: 'westend',
-      text: t('rpc.westend', 'Westend', { ns: 'apps-config' }),
+      info: 'vodka',
+      text: t('rpc.vodka', 'Vodka', { ns: 'apps-config' }),
       providers: {
-        Parity: 'wss://westend-rpc.polkadot.io',
-        'NodeFactory(Vedran)': 'wss://westend.vedran.nodefactory.io/ws'
+        Vodka: 'wss://vodka.rpc.neatcoin.org/ws'
+      }
+    },
+    {
+      info: 'web3games',
+      text: t('rpc.web3games', 'Web3Games', { ns: 'apps-config' }),
+      providers: {
+        Web3Games: 'wss://substrate.org.cn:4443'
       }
     },
     {
@@ -296,28 +348,6 @@ export function createTesting (t: TFunction): LinkOption[] {
       text: t('rpc.zero', 'Zero', { ns: 'apps-config' }),
       providers: {
         ZERO: 'wss://alphaville.zero.io'
-      }
-    },
-    {
-      info: 'mybank',
-      text: t('rpc.mybank', 'mybank.network', { ns: 'apps-config' }),
-      providers: {
-        MYBANK: 'wss://app.mybank.network/substrate'
-      }
-    },
-    {
-      info: 'unitv',
-      text: t('rpc.unitv', 'Unit Network', { ns: 'apps-config' }),
-      providers: {
-        'Unit Network': 'wss://unitventures.io/'
-      }
-    },
-    {
-      info: 'nftmart',
-      text: t('rpc.nftmart', 'Nftmart', { ns: 'apps-config' }),
-      providers: {
-        NftmartDev: 'ws://8.136.111.191:9944',
-        NftmartStaging: 'ws://82.157.37.77:9944'
       }
     }
   ]);