+++ packages/apps-config/src/ui/colors.ts
@@ -19,12 +19,15 @@ const chainChainx = '#F6C94A';
 const chainDarwinia = 'linear-gradient(-45deg, #FE3876 0%, #7C30DD 71%, #3A30DD 100%)';
 const chainDotMog = '#020609';
 const chainGalital = '#00063F';
+const chainGamePower = '#5d21a5';
 const chainHanonycash = '#0099CC';
 const chainHydrate = '#000000';
 const chainIntegritee = '#15bcff';
+const chainIpse = '#08a1e8';
 const chainKulupu = '#003366';
 const chainManta = '#2070a6';
-const chainNftmart = '#815287';
+const chainMoonrock = '#3d1d5a';
+const chainNFTMart = '#815287';
 const chainParami = '#ee06e2';
 const chainPhala = '#a7e300';
 const chainPhoenix = '#d42181';
@@ -35,6 +38,7 @@ const chainPolkadot = '#e6007a';
 const chainPolkaFoundry = '#ff527c';
 const chainPrism = 'linear-gradient(45deg, rgba(63,94,251,1) 0%, rgba(252,70,107,1) 100%)';
 const chainKilt = '#8c175b';
+const chainKonomi = '#007aff';
 const chainKusama = '#000000';
 const chainLitentry = 'linear-gradient(45deg, #5cc27c 0%, #6de98f 100%)';
 const chainMybank = '#282736';
@@ -50,9 +54,7 @@ const chainRoccoTick = '#22bb22';
 const chainRoccoTrack = '#bb2222';
 const chainRoccoTrick = '#2222bb';
 const chainRiochain = '#4d87f6';
-const chainSgc = 'linear-gradient(45deg, #0099F7 0%, #F11712 100%)';
 const chainSnakenet = '#f653a2';
-const chainSunrock = '#ffd13b';
 const chainWestend = '#da68a7';
 const chainGalois = '#000000';
 const chainZero = '#000000';
@@ -66,6 +68,7 @@ const chainSubsocial = '#b9018c';
 const chainTrustBase = '#ff43aa';
 const chainUnitv = '#1452F0';
 const chainVln = '#33cc33';
+const chainWeb3games = 'linear-gradient(45deg, #0099F7 0%, #F11712 100%)';
 const chainWestlake = 'linear-gradient(-90deg, #9400D3 0%, #5A5CA9 50%, #00BFFF 100%)';
 // based on node name
 // alphabetical
@@ -80,26 +83,31 @@ const nodeEncointerNotee = '#cc0000';
 const nodeEncointerTeeproxy = '#0000cc';
 const nodeEquilibrium = '#1792ff';
 const nodeGalital = '#00063F;';
+const nodeGamePower = '#5d21a5';
+const nodeIpse = '#08a1e8';
 const nodeJupiter = '#7143ff';
+const nodeKonomi = '#007aff';
 const nodeLitentry = 'linear-gradient(45deg, #5cc27c 0%, #6de98f 100%)';
 const nodeManta = '#2070a6';
 const nodeMoonbeam = '#53cbc9';
 const nodeMybank = '#282736';
-const nodeNftmart = '#307182';
+const nodeNFTMart = '#307182';
 const nodeNodle = '#1ab394';
 const nodeParami = '#ee06e2';
 const nodePolkadex = '#7C30DD';
+const nodePolymesh = '#1348e4';
 const nodePrism = 'linear-gradient(45deg, rgba(63,94,251,1) 0%, rgba(252,70,107,1) 100%)';
-const nodeRealis = 'linear-gradient(45deg, #6CFBE9 0%, #CFB241 100%)';
+const nodeRealis = 'linear-gradient(45deg, #E8AAC9 0%, #C4D9E7 50%, #EFD6E0 100%)';
 const nodeRiochain = '#1A3BB3';
-const nodeSgc = 'linear-gradient(45deg, #0099F7 0%, #F11712 100%)';
 const nodeSora = '#2D2926';
 const nodeStafi = '#00F3AB';
+const nodeStatemint = '#86e62a'; // '#98ff98';
 const nodeSubDAO = 'linear-gradient(50deg, #F20092 0%, #FF4D5D 100%)';
 const nodeSubsocial = '#b9018c';
 const nodeTernoa = '#d622ff';
 const nodeUniarts = chainUniarts;
 const nodeUnique = chainUnique;
+const nodeWeb3games = 'linear-gradient(45deg, #0099F7 0%, #F11712 100%)';
 const nodeWestlake = chainWestlake;
 const nodeZeitgeist = 'linear-gradient(180deg, rgba(32,90,172,1) 0%, rgba(26,72,138,1) 50%, rgba(13,36,69,1) 100%)';
 const nodeZero = '#0099cc';
@@ -130,16 +138,21 @@ export const chainColors: Record<string, any> = [
   ['Encointer PC1', nodeEncointerNotee],
   ['Galital', chainGalital],
   ['Galois', chainGalois],
+  ['GamePower Network', chainGamePower],
+  ['Halongbay PC1', chainPolkaFoundry],
   ['hanonycash', chainHanonycash],
   ['HydraDX Hydrate', chainHydrate],
   ['HydraDX Snakenet', chainSnakenet],
   ['HydraDX Snakenet Gen2', chainSnakenet],
+  ['HydraDX Snakenet Gen3', chainSnakenet],
   ['Idavoll', chainIdavoll],
   ['IntegriTEE PC1', chainIntegritee],
+  ['IpseTestnet', chainIpse],
   ['Jupiter A1', chainJupiter],
   ['Jupiter PC1', chainJupiter],
   ['KILT Collator Rococo', chainKilt],
   ['KILT Testnet', chainKilt],
+  ['Konomi', chainKonomi],
   ['Kulupu', chainKulupu],
   ['Kusama CC1', chainKusama],
   ['Kusama CC2', chainKusama],
@@ -152,12 +165,14 @@ export const chainColors: Record<string, any> = [
   ['Moonbase Alpha', nodeMoonbeam],
   ['Moonbase Stage', nodeMoonbeam],
   ['Moonbase Development Testnet', nodeMoonbeam],
+  ['Moonrock', chainMoonrock],
   ['mybank.network Testnet', chainMybank],
-  ['Nftmart Testnet', chainNftmart],
-  ['Nftmart Staging', chainNftmart],
+  ['NFTMart Testnet', chainNFTMart],
+  ['NFTMart Staging', chainNFTMart],
   ['Parami PC2', chainParami],
   ['Phala PC1', chainPhala],
-  ['Phala poc 3', chainPhala],
+  ['Phala PoC 4', chainPhala],
+  ['PHOENIX', chainPhoenix],
   ['Plasm', chainPlasm],
   ['Plasm PC2', chainPlasm],
   ['PolkaBTC', chainPolkabtc],
@@ -167,28 +182,27 @@ export const chainColors: Record<string, any> = [
   ['Polkadot', chainPolkadot],
   ['Prism Testnet', chainPrism],
   ['Prism PC1', chainPrism],
-  ['Halongbay Testnet', chainPolkaFoundry],
   ['ReAlis Network', nodeRealis],
   ['Rococo', chainRocco],
   ['RioChain CC-1', chainRiochain],
   ['Riochain Staging', chainRiochain],
-  ['Sgc ', chainSgc],
+  ['Statemint Test', nodeStatemint],
   ['SubDAO PC1', chainSubDAO],
   ['Subsocial PC1', chainSubsocial],
-  ['Sunrock', chainSunrock],
   ['Tick', chainRoccoTick],
   ['Track', chainRoccoTrack],
   ['Trick', chainRoccoTrick],
   ['TrustBase PC1', chainTrustBase],
   ['uni arts staging network', chainUniarts],
+  ['UniArts Mainnet', chainUniarts],
   ['Unique Node', chainUnique],
+  ['UNIT', chainUnitv],
   ['VLN PC', chainVln],
+  ['Web3games ', chainWeb3games],
   ['Westend', chainWestend],
   ['Westlake', chainWestlake],
   ['Zenlink PC1', chainZenlink],
-  ['ZERO.IO', chainZero],
-  ['PHOENIX', chainPhoenix],
-  ['UNIT', chainUnitv]
+  ['ZERO.IO', chainZero]
 ].reduce((colors, [chain, color]): Record<string, any> => ({
   ...colors,
   [chain.toLowerCase()]: color
@@ -209,37 +223,41 @@ export const nodeColors: Record<string, any> = [
   ['Encointer Node noTEE', nodeEncointerNotee],
   ['Encointer Node TEE proxy', nodeEncointerTeeproxy],
   ['Equilibrium node', nodeEquilibrium],
+  ['Galital', nodeGalital],
   ['Galital Parachain Collator', nodeGalital],
+  ['GamePower Node', nodeGamePower],
+  ['IpseTestnet', nodeIpse],
+  ['Konomi Collator', nodeKonomi],
   ['Litentry Collator', nodeLitentry],
   ['Manta Collator', nodeManta],
   ['mybank.network node', nodeMybank],
-  ['Nftmart Testnet', nodeNftmart],
-  ['Nftmart Staging', nodeNftmart],
+  ['NFTMart Testnet', nodeNFTMart],
+  ['NFTMart Staging', nodeNFTMart],
   ['nodle chain node', nodeNodle],
   ['Parami Collator', nodeParami],
   ['Patract Node', nodeJupiter],
   ['Polkadex Node', nodePolkadex],
+  ['Polymesh Node', nodePolymesh],
   ['Prism Node', nodePrism],
   ['Prism Collator', nodePrism],
   ['ReAlis Network', nodeRealis],
   ['Rio Defi Chain Node', nodeRiochain],
   ['Riochain Staging', nodeRiochain],
-  ['Sgc', nodeSgc],
   ['SORA', nodeSora],
   ['Stafi node', nodeStafi],
+  ['Statemint Collator', nodeStatemint],
+  ['SubDAO Collator', nodeSubDAO],
   ['subsocial node', nodeSubsocial],
   ['subzero node', nodeZero],
   ['Ternoa Node', nodeTernoa],
   ['uni arts node', nodeUniarts],
+  ['UniArts Node', nodeUniarts],
   ['Unique Node', nodeUnique],
+  ['Web3games', nodeWeb3games],
   ['Westlake', nodeWestlake],
   ['Zeitgeist Node', nodeZeitgeist],
   ['Zeitgeist Collator', nodeZeitgeist],
-  ['Zenlink Collator', nodeZenlink],
-  ['SubDAO Collator', nodeSubDAO]
-  // ['node template', emptyColor],
-  // ['parity polkadot', emptyColor],
-  // ['substrate node', emptyColor]
+  ['Zenlink Collator', nodeZenlink]
 ].reduce((colors, [node, color]): Record<string, any> => ({
   ...colors,
   [node.toLowerCase().replace(/-/g, ' ')]: color